Pytest_Framework/requirements.txt
Appium-Python-Client==2.10.1
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
et-xmlfile==1.1.0
exceptiongroup==1.1.1
execnet==1.9.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PySocks==1.7.1
pytest==7.4.0
pytest-catchlog==1.2.2
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-xdist==3.3.1
requests==2.31.0
selenium==4.9.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
urllib3>=1.21.1,<3
wsproto==1.2.0